addressing more pr comments
