+++ web-client/src/applicationContext.js
@@ -72,8 +72,8 @@ import { getItem } from '../../shared/src/persistence/localStorage/getItem';
 import { getItemInteractor } from '../../shared/src/business/useCases/getItemInteractor';
 import { getNotificationsInteractor } from '../../shared/src/proxies/users/getNotificationsProxy';
 import { getProcedureTypesInteractor } from '../../shared/src/business/useCases/getProcedureTypesInteractor';
-import { getScannerInterface } from '../../shared/src/business/useCases/getScannerInterfaceInteractor';
-import { getScannerInterface as getScannerMockInterfaceInteractor } from '../../shared/src/business/useCases/getScannerMockInterfaceInteractor';
+import { getScannerInterface } from '../../shared/src/persistence/dynamsoft/getScannerInterface';
+import { getScannerInterface as getScannerMockInterfaceInteractor } from '../../shared/src/persistence/dynamsoft/getScannerMockInterface';
 import { getSentMessagesForSectionInteractor } from '../../shared/src/proxies/workitems/getSentMessagesForSectionProxy';
 import { getSentMessagesForUserInteractor } from '../../shared/src/proxies/workitems/getSentMessagesForUserProxy';
 import { getTrialSessionDetailsInteractor } from '../../shared/src/proxies/trialSessions/getTrialSessionDetailsProxy';
